Minor fixes to get the client past few validation errors.
We turned off validation on the server side so these fixes might not be needed
in that case. Also, there are probably other validations issues with the client
code. With the server not doing validation on incoming resources, it would be
useful to add a $validate call after the client posts a resource. That way we
both get the resource to the server (even if invalid) but also call to get ALL
the validation issues at once to review and fix instead of the resource being
rejected for each validation error, one at a time.